ci: move validtion files to utils
